Replace QuantizedFunctionsCUDA with Triton version
