+++ nncf/torch/quantization/quantize_functions.py
@@ -19,7 +19,7 @@ from nncf.errors import ValidationError
 from nncf.torch.dynamic_graph.patch_pytorch import register_operator
 from nncf.torch.quantization.extensions import QuantizedFunctionsCPU
 from nncf.torch.quantization.reference import ReferenceQuantizedFunctions as RQ
-from nncf.torch.quantization.triton.reference import TritonQuantizedFunctions as TQ
+from nncf.torch.quantization.triton import QuantizedFunctionsCUDA
 from nncf.torch.utils import add_domain
 
 
@@ -38,7 +38,7 @@ class QuantizeSymmetric(torch.autograd.Function):
             if not input_.is_contiguous():
                 nncf_logger.debug("input_ is not contiguous!")
                 input_ = input_.contiguous()
-            output = TQ.Quantize_forward(input_, input_low, input_range, levels)
+            output = QuantizedFunctionsCUDA.get("Quantize_forward")(input_, input_low, input_range, levels)
         else:
             output = QuantizedFunctionsCPU.get("Quantize_forward")(input_, input_low, input_range, levels)
 
@@ -62,7 +62,7 @@ class QuantizeSymmetric(torch.autograd.Function):
                 nncf_logger.debug("grad_output is not contiguous!")
                 grad_output = grad_output.contiguous()
 
-            grad_input, _, grad_scale = TQ.Quantize_backward(
+            grad_input, _, grad_scale = QuantizedFunctionsCUDA.get("Quantize_backward")(
                 grad_output, input_, input_low, input_range, levels, level_low, level_high
             )
         else:
@@ -86,7 +86,7 @@ class QuantizeAsymmetric(torch.autograd.Function):
             if not input_.is_contiguous():
                 nncf_logger.debug("input_ is not contiguous!")
                 input_ = input_.contiguous()
-            output = TQ.Quantize_forward(input_, input_low, input_range, levels)
+            output = QuantizedFunctionsCUDA.get("Quantize_forward")(input_, input_low, input_range, levels)
         else:
             output = QuantizedFunctionsCPU.get("Quantize_forward")(input_, input_low, input_range, levels)
 
@@ -109,7 +109,7 @@ class QuantizeAsymmetric(torch.autograd.Function):
                 nncf_logger.debug("grad_output is not contiguous!")
                 grad_output = grad_output.contiguous()
 
-            grad_input, grad_input_low, grad_input_range = TQ.Quantize_backward(
+            grad_input, grad_input_low, grad_input_range = QuantizedFunctionsCUDA.get("Quantize_backward")(
                 grad_output, input_, input_low, input_range, levels, level_low, level_high
             )
         else:
@@ -131,7 +131,7 @@ class QuantizeSymmetricTorch(torch.autograd.Function):
         input_ = input_.reshape(input_shape)
 
         if input_.is_cuda:
-            output = TQ.Quantize_forward(input_, input_low, input_range, levels)
+            output = QuantizedFunctionsCUDA.get("Quantize_forward")(input_, input_low, input_range, levels)
         else:
             output = RQ.Quantize_forward(input_.type(torch.float32), input_low, input_range, levels)
 
@@ -155,7 +155,7 @@ class QuantizeSymmetricTorch(torch.autograd.Function):
         grad_output = grad_output.reshape(input_shape)
 
         if input_.is_cuda:
-            grad_input, _, grad_scale = TQ.Quantize_backward(
+            grad_input, _, grad_scale = QuantizedFunctionsCUDA.get("Quantize_backward")(
                 grad_output, input_, input_low, input_range, levels, level_low, level_high
             )
         else:
@@ -177,7 +177,7 @@ class QuantizeAsymmetricTorch(torch.autograd.Function):
         input_ = input_.reshape(input_shape)
 
         if input_.is_cuda:
-            output = TQ.Quantize_forward(input_, input_low, input_range, levels)
+            output = QuantizedFunctionsCUDA.get("Quantize_forward")(input_, input_low, input_range, levels)
         else:
             output = RQ.Quantize_forward(input_.type(torch.float32), input_low, input_range, levels)
 
@@ -201,7 +201,7 @@ class QuantizeAsymmetricTorch(torch.autograd.Function):
         grad_output = grad_output.reshape(input_shape)
 
         if input_.is_cuda:
-            grad_input, grad_low, grad_range = TQ.Quantize_backward(
+            grad_input, grad_low, grad_range = QuantizedFunctionsCUDA.get("Quantize_backward")(
                 grad_output, input_, input_low, input_range, levels, level_low, level_high
             )
         else:
